Pulled the part of the gs function which does random mating out and put it as a seperate internal function, since we'll likely use that to generate new genomes for unsampled individuals when doing simulations.
Also, saved one example of the JWAS inputs and outputs to the r01 folder for Hao to take a look at. ie are the input phenotypes, ig are the input genotypes for JWAS. For this run, the prior residual variance given to JWAS was 106.643986264865 and the prior genetic variance was 123.185698013811 (a random adjustment of up to +-25%). True heritability was 0.5.